+++ test/WebJobs.Script.Tests/Rpc/RpcWorkerConfigTests.cs
@@ -237,220 +237,6 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Rpc
             }
         }
 
-        [Theory]
-        [InlineData("%FUNCTIONS_WORKER_RUNTIME_VERSION%/{os}/{architecture}", "3.7/LINUX/X64")]
-        [InlineData("%FUNCTIONS_WORKER_RUNTIME_VERSION%/{architecture}", "3.7/X64")]
-        [InlineData("%FUNCTIONS_WORKER_RUNTIME_VERSION%/{os}", "3.7/LINUX")]
-        public void LanguageWorker_HydratedWorkerPath_EnvironmentVersionSet(string defaultWorkerPath, string expectedPath)
-        {
-            _testEnvironment.SetEnvironmentVariable(LanguageWorkerConstants.FunctionWorkerRuntimeVersionSettingName, "3.7");
-
-            RpcWorkerDescription workerDescription = new RpcWorkerDescription()
-            {
-                Arguments = new List<string>(),
-                DefaultExecutablePath = "python",
-                DefaultWorkerPath = defaultWorkerPath,
-                DefaultRuntimeVersion = "3.6",
-                SupportedArchitectures = new List<string>() { Architecture.X64.ToString(), Architecture.X86.ToString() },
-                SupportedRuntimeVersions = new List<string>() { "3.6", "3.7" },
-                SupportedOperatingSystems = new List<string>()
-                    {
-                        OSPlatform.Windows.ToString(),
-                        OSPlatform.OSX.ToString(),
-                        OSPlatform.Linux.ToString()
-                    },
-                WorkerDirectory = string.Empty,
-                Extensions = new List<string>() { ".py" },
-                Language = "python"
-            };
-            var configBuilder = ScriptSettingsManager.CreateDefaultConfigurationBuilder()
-                  .AddInMemoryCollection(new Dictionary<string, string>
-                  {
-                      ["languageWorker"] = "test"
-                  });
-            var config = configBuilder.Build();
-            var scriptSettingsManager = new ScriptSettingsManager(config);
-            var testLogger = new TestLogger("test");
-            var configFactory = new WorkerConfigFactory(config, testLogger, _testSysRuntimeInfo, _testEnvironment);
-
-            Assert.Equal(expectedPath, configFactory.GetHydratedWorkerPath(workerDescription));
-        }
-
-        [Theory]
-        [InlineData("%FUNCTIONS_WORKER_RUNTIME_VERSION%/{os}/{architecture}", "3.6/LINUX/X64")]
-        [InlineData("%FUNCTIONS_WORKER_RUNTIME_VERSION%/{architecture}", "3.6/X64")]
-        [InlineData("%FUNCTIONS_WORKER_RUNTIME_VERSION%/{os}", "3.6/LINUX")]
-        public void LanguageWorker_HydratedWorkerPath_EnvironmentVersionNotSet(string defaultWorkerPath, string expectedPath)
-        {
-            // We fall back to the default version when this is not set
-            // Environment.SetEnvironmentVariable(LanguageWorkerConstants.FunctionWorkerRuntimeVersionSettingName, "3.7");
-
-            RpcWorkerDescription workerDescription = new RpcWorkerDescription()
-            {
-                Arguments = new List<string>(),
-                DefaultExecutablePath = "python",
-                SupportedArchitectures = new List<string>() { Architecture.X64.ToString(), Architecture.X86.ToString() },
-                SupportedRuntimeVersions = new List<string>() { "3.6", "3.7" },
-                SupportedOperatingSystems = new List<string>()
-                    {
-                        OSPlatform.Windows.ToString(),
-                        OSPlatform.OSX.ToString(),
-                        OSPlatform.Linux.ToString()
-                    },
-                DefaultWorkerPath = defaultWorkerPath,
-                WorkerDirectory = string.Empty,
-                Extensions = new List<string>() { ".py" },
-                Language = "python",
-                DefaultRuntimeVersion = "3.6"
-            };
-            var configBuilder = ScriptSettingsManager.CreateDefaultConfigurationBuilder()
-                  .AddInMemoryCollection(new Dictionary<string, string>
-                  {
-                      ["languageWorker"] = "test"
-                  });
-            var config = configBuilder.Build();
-            var scriptSettingsManager = new ScriptSettingsManager(config);
-            var testLogger = new TestLogger("test");
-            var configFactory = new WorkerConfigFactory(config, testLogger, _testSysRuntimeInfo, _testEnvironment);
-
-            Assert.Equal(expectedPath, configFactory.GetHydratedWorkerPath(workerDescription));
-        }
-
-        [Theory]
-        [InlineData(Architecture.Arm)]
-        [InlineData(Architecture.Arm64)]
-        public void LanguageWorker_HydratedWorkerPath_UnsupportedArchitecture(Architecture unsupportedArch)
-        {
-            RpcWorkerDescription workerDescription = new RpcWorkerDescription()
-            {
-                Arguments = new List<string>(),
-                DefaultExecutablePath = "python",
-                DefaultWorkerPath = "{architecture}/worker.py",
-                WorkerDirectory = string.Empty,
-                SupportedArchitectures = new List<string>() { Architecture.X64.ToString(), Architecture.X86.ToString() },
-                SupportedRuntimeVersions = new List<string>() { "3.6", "3.7" },
-                SupportedOperatingSystems = new List<string>()
-                    {
-                        OSPlatform.Windows.ToString(),
-                        OSPlatform.OSX.ToString(),
-                        OSPlatform.Linux.ToString()
-                    },
-                Extensions = new List<string>() { ".py" },
-                Language = "python",
-                DefaultRuntimeVersion = "3.7"
-            };
-            var configBuilder = ScriptSettingsManager.CreateDefaultConfigurationBuilder()
-                  .AddInMemoryCollection(new Dictionary<string, string>
-                  {
-                      ["languageWorker"] = "test"
-                  });
-            var config = configBuilder.Build();
-            var scriptSettingsManager = new ScriptSettingsManager(config);
-            var testLogger = new TestLogger("test");
-            Mock<ISystemRuntimeInformation> mockRuntimeInfo = new Mock<ISystemRuntimeInformation>();
-            mockRuntimeInfo.Setup(r => r.GetOSArchitecture()).Returns(unsupportedArch);
-            mockRuntimeInfo.Setup(r => r.GetOSPlatform()).Returns(OSPlatform.Linux);
-            var configFactory = new WorkerConfigFactory(config, testLogger, mockRuntimeInfo.Object, _testEnvironment);
-
-            var ex = Assert.Throws<PlatformNotSupportedException>(() => configFactory.GetHydratedWorkerPath(workerDescription));
-            Assert.Equal(ex.Message, $"Architecture {unsupportedArch.ToString()} is not supported for language {workerDescription.Language}");
-        }
-
-        [Fact]
-        public void LanguageWorker_HydratedWorkerPath_UnsupportedOS()
-        {
-            OSPlatform bogusOS = OSPlatform.Create("BogusOS");
-            RpcWorkerDescription workerDescription = new RpcWorkerDescription()
-            {
-                Arguments = new List<string>(),
-                DefaultExecutablePath = "python",
-                DefaultWorkerPath = "{os}/worker.py",
-                SupportedOperatingSystems = new List<string>()
-                    {
-                        OSPlatform.Windows.ToString(),
-                        OSPlatform.OSX.ToString(),
-                        OSPlatform.Linux.ToString()
-                    },
-                WorkerDirectory = string.Empty,
-                Extensions = new List<string>() { ".py" },
-                Language = "python",
-                DefaultRuntimeVersion = "3.7"
-            };
-            var configBuilder = ScriptSettingsManager.CreateDefaultConfigurationBuilder()
-                  .AddInMemoryCollection(new Dictionary<string, string>
-                  {
-                      ["languageWorker"] = "test"
-                  });
-            var config = configBuilder.Build();
-            var scriptSettingsManager = new ScriptSettingsManager(config);
-            var testLogger = new TestLogger("test");
-            Mock<ISystemRuntimeInformation> mockRuntimeInfo = new Mock<ISystemRuntimeInformation>();
-            mockRuntimeInfo.Setup(r => r.GetOSArchitecture()).Returns(Architecture.X64);
-            mockRuntimeInfo.Setup(r => r.GetOSPlatform()).Returns(bogusOS);
-            var configFactory = new WorkerConfigFactory(config, testLogger, mockRuntimeInfo.Object, _testEnvironment);
-
-            var ex = Assert.Throws<PlatformNotSupportedException>(() => configFactory.GetHydratedWorkerPath(workerDescription));
-            Assert.Equal(ex.Message, $"OS BogusOS is not supported for language {workerDescription.Language}");
-        }
-
-        [Fact]
-        public void LanguageWorker_HydratedWorkerPath_UnsupportedDefaultRuntimeVersion()
-        {
-            RpcWorkerDescription workerDescription = new RpcWorkerDescription()
-            {
-                Arguments = new List<string>(),
-                DefaultExecutablePath = "python",
-                SupportedRuntimeVersions = new List<string>() { "3.6", "3.7" },
-                DefaultWorkerPath = $"{LanguageWorkerConstants.RuntimeVersionPlaceholder}/worker.py",
-                WorkerDirectory = string.Empty,
-                Extensions = new List<string>() { ".py" },
-                Language = "python",
-                DefaultRuntimeVersion = "3.5"
-            };
-            var configBuilder = ScriptSettingsManager.CreateDefaultConfigurationBuilder()
-                  .AddInMemoryCollection(new Dictionary<string, string>
-                  {
-                      ["languageWorker"] = "test"
-                  });
-            var config = configBuilder.Build();
-            var scriptSettingsManager = new ScriptSettingsManager(config);
-            var testLogger = new TestLogger("test");
-            var configFactory = new WorkerConfigFactory(config, testLogger, _testSysRuntimeInfo, _testEnvironment);
-
-            var ex = Assert.Throws<NotSupportedException>(() => configFactory.GetHydratedWorkerPath(workerDescription));
-            Assert.Equal(ex.Message, $"Version {workerDescription.DefaultRuntimeVersion} is not supported for language {workerDescription.Language}");
-        }
-
-        [Fact]
-        public void LanguageWorker_HydratedWorkerPath_UnsupportedEnvironmentRuntimeVersion()
-        {
-            _testEnvironment.SetEnvironmentVariable(LanguageWorkerConstants.FunctionWorkerRuntimeVersionSettingName, "3.4");
-
-            RpcWorkerDescription workerDescription = new RpcWorkerDescription()
-            {
-                Arguments = new List<string>(),
-                DefaultExecutablePath = "python",
-                SupportedRuntimeVersions = new List<string>() { "3.6", "3.7" },
-                DefaultWorkerPath = $"{LanguageWorkerConstants.RuntimeVersionPlaceholder}/worker.py",
-                WorkerDirectory = string.Empty,
-                Extensions = new List<string>() { ".py" },
-                Language = "python",
-                DefaultRuntimeVersion = "3.7" // Ignore this if environment is set
-            };
-            var configBuilder = ScriptSettingsManager.CreateDefaultConfigurationBuilder()
-                  .AddInMemoryCollection(new Dictionary<string, string>
-                  {
-                      ["languageWorker"] = "test"
-                  });
-            var config = configBuilder.Build();
-            var scriptSettingsManager = new ScriptSettingsManager(config);
-            var testLogger = new TestLogger("test");
-            var configFactory = new WorkerConfigFactory(config, testLogger, _testSysRuntimeInfo, _testEnvironment);
-
-            var ex = Assert.Throws<NotSupportedException>(() => configFactory.GetHydratedWorkerPath(workerDescription));
-            Assert.Equal(ex.Message, $"Version 3.4 is not supported for language {workerDescription.Language}");
-        }
-
         private IEnumerable<WorkerConfig> TestReadWorkerProviderFromConfig(IEnumerable<TestLanguageWorkerConfig> configs, ILogger testLogger, string language = null, Dictionary<string, string> keyValuePairs = null, bool appSvcEnv = false)
         {
             Mock<IEnvironment> mockEnvironment = new Mock<IEnvironment>();
